refactor(zod-v3): extract slim-primitives walker to its own file
Mirror v4's file layout (zod-v4/slim-primitives.ts). The walker is
self-contained — only depends on isZodSchemaType + SlimPrimitiveKind
— so it lifts cleanly out of the monolith without touching any
caller. PERMISSIVE_V3 is co-located with the walker since it's only
referenced by it.

The remaining strip / stripRoot / getSlimSchema chunk in index.ts has
cross-cutting dependencies (unwrapToDiscriminatedUnion lives in
index.ts) that make a clean extraction harder. Deferred — it would
require reorganising more of the monolith for marginal tree-shake
benefit.

  - new src/runtime/adapters/zod-v3/slim-primitives.ts (~165 LOC moved)
  - index.ts down ~165 LOC; imports the extracted walker
